控制

+++ camera_try/user/contorl/Smart_key.c
@@ -9,7 +9,7 @@
 #include "Smart_headfile.h"
 
 int16 Key_Val,Key_Down,Key_Up,Key_Old;
-uint8 key_slow = 0;
+int16 key_slow = 0;
 
 unsigned char Key_Read(void)
 {
