+++ camera_try/user/contorl/Smart_key.h
@@ -11,7 +11,7 @@
 #include "zf_common_headfile.h"
 
 extern int16 Key_Val,Key_Down,Key_Up,Key_Old;
-extern uint8 key_slow ;
+extern int16 key_slow ;
 
 unsigned char Key_Read(void);
 void Key_Proc(int16 *Key_Val,int16 *Key_Old,int16 *Key_Down,int16 *Key_Up);
